+++ src/showI18n.js
@@ -5,8 +5,9 @@ const { operation } = require('./utils/constant');
 module.exports = (context) => {
 	context.subscriptions.push(
 		registerCommand(operation.showI18n.cmd, (uri) => {
-			if (uri && uri.path) {
-				openFileByPath(uri.path).then((editor) => {
+			const fsPath = uri && (uri.fsPath || uri.path);
+			if (fsPath) {
+				openFileByPath(fsPath).then((editor) => {
 					showJson({ editor, context });
 				});
 			} else {